* akka version 2.5.13 * Roach serialization moved to separated class
