+++ akka-service-commons/src/main/scala/org/s4s0l/betelgeuse/akkacommons/persistence/crate/CrateDbLocks.scala
@@ -20,6 +20,7 @@ import java.sql.Timestamp
 import java.util.{Calendar, Date, UUID}
 
 import akka.actor.{Cancellable, Scheduler}
+import akka.stream.scaladsl.Source
 import com.typesafe.config.Config
 import io.crate.shade.org.postgresql.util.PSQLException
 import org.s4s0l.betelgeuse.akkacommons.persistence.utils.DbLocksSettings.{DbLocksRolling, DbLocksSingle}
@@ -106,11 +107,14 @@ class CrateDbLocks(val schema: String = "locks", locksTable: String = "locks")
       case _: DbLocksSingle => runLockedInternal()
       case DbLocksRolling(duration, _, _, lockTimeInitial) => runLockedInternal(
         Some(
-          scheduler.schedule(duration - lockTimeInitial, duration, () =>
-            if (!isFinished) {
-              lock(lockName, txExecutor, settings)
-            }))
-      )
+          scheduler.schedule(duration - lockTimeInitial, duration, new Runnable {
+            override def run(): Unit = {
+              if (!isFinished) {
+                lock(lockName, txExecutor, settings)
+              }
+            }
+          })
+        ))
     }
 
   }
@@ -178,6 +182,7 @@ class CrateDbLocks(val schema: String = "locks", locksTable: String = "locks")
       tryNTimesExceptionFactory(s"Taking lock failed. Holder $uuid"),
       (lockSettings.lockAttemptInterval / 2).toMillis) {
       txExecutor.doInTx { implicit session =>
+        Source.combine()
         lockAttempt(lockName, lockSettings)(session)
       }
     }
