+++ gradle/akka.gradle
@@ -16,7 +16,7 @@
 
 ext {
     akkaBinaryVersion = '2.5'
-    akkaVersion = '2.5.11'
+    akkaVersion = '2.5.13'
     akkaHttpVersion = '10.0.10'
 }
